Move SettingsUseSnippetSwitch to settings.
