+++ assets/js/modules/analytics-4/components/common/index.js
@@ -17,5 +17,4 @@
  */
 
 export { default as PropertySelect } from './PropertySelect';
-export { default as SettingsUseSnippetSwitch } from './SettingsUseSnippetSwitch';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
