+++ assets/js/modules/tagmanager/components/settings/SettingsUseSnippetSwitch.js
@@ -1,5 +1,5 @@
 /**
- * Analytics Settings Use Snippet Switch component.
+ * Tag Manager Settings Use Snippet Switch component.
  *
  * Site Kit by Google, Copyright 2022 Google LLC
  *
@@ -26,55 +26,34 @@ import { __, sprintf } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { MODULES_ANALYTICS } from '../../datastore/constants';
-import UseSnippetSwitch from './UseSnippetSwitch';
+import UseSnippetSwitch from '../common/UseSnippetSwitch';
+import { MODULES_TAGMANAGER } from '../../datastore/constants';
 const { useSelect } = Data;
 
 export default function SettingsUseSnippetSwitch() {
 	const useSnippet = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getUseSnippet()
+		select( MODULES_TAGMANAGER ).getUseSnippet()
 	);
-	const canUseSnippet = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getCanUseSnippet()
+
+	const primaryContainerID = useSelect( ( select ) =>
+		select( MODULES_TAGMANAGER ).getPrimaryContainerID()
 	);
 
 	const existingTag = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getExistingTag()
+		select( MODULES_TAGMANAGER ).getExistingTag()
 	);
-	const propertyID = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS ).getPropertyID()
-	);
-
-	if ( canUseSnippet === undefined ) {
-		return null;
-	}
-
-	if ( canUseSnippet === false ) {
-		return (
-			<UseSnippetSwitch
-				description={
-					<p>
-						{ __(
-							'The code is controlled by the Tag Manager module.',
-							'google-site-kit'
-						) }
-					</p>
-				}
-			/>
-		);
-	}
 
 	let description;
 
 	if ( existingTag ) {
 		description =
-			existingTag === propertyID ? (
+			primaryContainerID === existingTag ? (
 				<Fragment>
 					<p>
 						{ sprintf(
 							/* translators: %s: existing tag ID */
 							__(
-								'A tag %s for the selected property already exists on the site.',
+								'A tag %s for the selected container already exists on the site.',
 								'google-site-kit'
 							),
 							existingTag
@@ -82,7 +61,7 @@ export default function SettingsUseSnippetSwitch() {
 					</p>
 					<p>
 						{ __(
-							'Make sure you remove it if you want to place the same UA tag via Site Kit, otherwise they will be duplicated.',
+							'Make sure you remove it if you want to place the same tag via Site Kit, otherwise they will be duplicated.',
 							'google-site-kit'
 						) }
 					</p>
@@ -101,7 +80,7 @@ export default function SettingsUseSnippetSwitch() {
 					</p>
 					<p>
 						{ __(
-							'If you prefer to collect data using that existing UA tag, please select the corresponding account and property above.',
+							'If you prefer to collect data using that existing tag, please select the corresponding account and property above.',
 							'google-site-kit'
 						) }
 					</p>
@@ -111,14 +90,14 @@ export default function SettingsUseSnippetSwitch() {
 		description = useSnippet ? (
 			<p>
 				{ __(
-					'Site Kit will add the UA code automatically.',
+					'Site Kit will add the code automatically.',
 					'google-site-kit'
 				) }
 			</p>
 		) : (
 			<p>
 				{ __(
-					'Site Kit will not add the UA code to your site.',
+					'Site Kit will not add the code to your site.',
 					'google-site-kit'
 				) }
 			</p>
