+++ assets/js/modules/tagmanager/components/settings/index.js
@@ -18,4 +18,5 @@
 
 export { default as SettingsEdit } from './SettingsEdit';
 export { default as SettingsForm } from './SettingsForm';
+export { default as SettingsUseSnippetSwitch } from './SettingsUseSnippetSwitch';
 export { default as SettingsView } from './SettingsView';
